Week3 lab
Still confused about what numbers in the wine data set are input values
and output values and where the hidden layers are.

Somehow, i managed to define the network with the class identifier as
the output layer neuron number and number of other scores in a row as
the input layer neuron number.

Somehow, the hidden layers can be changed to any number...