Week 4 - part 01
Modify the ANN to obtain better results
How to make the training and test datasets similar?
Answer:
In order to improve the similarity between training and test data, it is
recommended to evenly represent all the categories of target values
(median value of owner-occupied homes) within each of training and test
dataset. For example, to random select the records into different
dataset, shuffling the whole data set may lead to a natural distribution
in training and test dataset. In addition, there is a recommended
strategy to generate random categorical data in training and test
dataset, respectively. Making up the training and test data sets is
based on the proportions of the different levels of target values
(different categories of homes median value) in the entire dataset. In
this way, evenly orgranized and similarly distributed dataset can be
obtained.