Assignment 6 Error
It seems like that when I just opened and ran the first and second file
in the notebook and there are error reports. I asked some other guy in
this class and we cannot identify what is wrong. And in the third file
after the improvement code, the output seems like that the code
functioned well but afterwards another error report came up. Would you
might help me identify where the problem is? Or is it that something is
wrong with my computer? Apologize and thanks!